Fix linting warnings and improve code quality
- Remove unused variables and imports across multiple files
- Fix Fast refresh warnings by separating constants from components
- Delete incomplete Navigation-backup.tsx file
- Improve OptimizedImage component with proper implementation
- Update ESLint configuration and fix remaining warnings
- All tests passing, code quality improved
