+++ app/components/NewsletterSignup.tsx
@@ -20,7 +20,7 @@ const NewsletterSignup: React.FC<NewsletterSignupProps> = ({ onSubscribe, classN
       }
       setMessage('Thank you for subscribing!');
       setEmail('');
-    } catch (error) {
+    } catch {
       setMessage('Something went wrong. Please try again.');
     } finally {
       setIsSubmitting(false);
