+++ app/types/next.d.ts
@@ -1,10 +1,7 @@
 import { NextPage } from 'next';
-import { AppProps } from 'next/app';
 import React from 'react';
 
-interface NextPageWithLayout<P = {}, IP = P> extends NextPage<P, IP> {
-  getLayout?: (_page: React.ReactElement) => React.ReactNode;
-}
+// NextPageWithLayout interface for layout support
 
 declare module 'next' {
   interface NextPageWithLayout<P = {}, IP = P> extends NextPage<P, IP> {
