+++ app/utils/accessibilityUtils.ts
@@ -1,5 +1,4 @@
 'use client';
-import React from 'react';
 import { useState, useEffect } from 'react';
 
 export const useAccessibilityUtils = () => {
