+++ app/utils/performance.ts
@@ -39,24 +39,24 @@ export class PerformanceMonitor {
     if (typeof window === "undefined") return;
 
     // Largest Contentful Paint
-    new PerformanceObserver((_entryList) => {
+    new PerformanceObserver((entryList) => {
       const entries = entryList.getEntries();
       const lastEntry = entries[entries.length - 1];
       this.metrics.set("LCP", lastEntry.startTime)}).observe({ entryTypes: ["largest-contentful-paint"] });
 
     // First Input Delay
-    new PerformanceObserver((_entryList) => {
+    new PerformanceObserver((entryList) => {
       const entries = entryList.getEntries();
-      entries.forEach((_entry) => {
+      entries.forEach((entry) => {
         // Use processingStart if available, otherwise calculate from startTime
         const processingStart = (entry as { processingStart?: number }).processingStart || entry.startTime;
         this.metrics.set("FID", processingStart - entry.startTime)})}).observe({ entryTypes: ["first-input"] });
 
     // Cumulative Layout Shift
     let clsValue = 0;
-    new PerformanceObserver((_entryList) => {
+    new PerformanceObserver((entryList) => {
       const entries = entryList.getEntries();
-      entries.forEach((_entry) => {
+      entries.forEach((entry) => {
         if (!(entry as { hadRecentInput?: boolean }).hadRecentInput) {
           clsValue += (entry as { value?: number }).value || 0}
       });
