+++ components/OptimizedImage.tsx
@@ -1,7 +1,24 @@
 'use client';
-import React from 'react';
 
 interface OptimizedImageProps {
-
-
+  src: string;
+  alt: string;
+  width?: number;
+  height?: number;
+  className?: string;
 }
+
+const OptimizedImage: React.FC<OptimizedImageProps> = ({ src, alt, width, height, className }) => {
+  return (
+    <img
+      src={src}
+      alt={alt}
+      width={width}
+      height={height}
+      className={className}
+      loading="lazy"
+    />
+  );
+};
+
+export default OptimizedImage;
